usage of error handler middleware
