+++ src/api/v2/middleware/errorHandler.middleware.js
@@ -1,0 +1,10 @@
+export const errorHandlerMiddleware = (err, req, res, next) => {
+  console.error(`Error: ${err.message}`);
+  console.error(err.stack);
+
+  res.status(err.statusCode || 500).json({
+    status: "error",
+    message: err.message || "Internal server error",
+    ...(process.env.NODE_ENV === "development" && { stack: err.stack }),
+  });
+};
