+++ src/api/v2/middleware/rateLimit.middleware.js
@@ -5,15 +5,3 @@ export const apiRateLimiter = rateLimit({
   max: 100, // limit each IP to 100 requests per windowMs
   message: "Too many requests, please try again later.",
 });
-
-// middleware/errorHandler.middleware.js
-export const errorHandlerMiddleware = (err, req, res, next) => {
-  console.error(`Error: ${err.message}`);
-  console.error(err.stack);
-
-  res.status(err.statusCode || 500).json({
-    status: "error",
-    message: err.message || "Internal server error",
-    ...(process.env.NODE_ENV === "development" && { stack: err.stack }),
-  });
-};
