+++ src/api/v2/routes/hianime.routes.js
@@ -3,20 +3,28 @@ import { cacheMiddleware } from "../../middleware/cache.middleware.js";
 import * as searchController from "../controllers/HiAnime/search.controller.js";
 import * as streamController from "../controllers/HiAnime/streamInfo.controller.js";
 import * as episodeListController from "../controllers/HiAnime/episodeList.controller.js";
+import { errorHandlerMiddleware } from "../middleware/errorHandler.middleware.js";
 
 const router = Router();
 
 router.get(
   "/episodes/:id",
   cacheMiddleware("episodeList"),
+  errorHandlerMiddleware,
   episodeListController.getEpisodes
 );
 
-router.get("/search", cacheMiddleware("search"), searchController.search);
+router.get(
+  "/search",
+  cacheMiddleware("search"),
+  errorHandlerMiddleware,
+  searchController.search
+);
 
 router.get(
   "/stream/:id",
   cacheMiddleware("streamInfo"),
+  errorHandlerMiddleware,
   streamController.getStreamInfo
 );
