+++ src/server.js
@@ -4,15 +4,17 @@ import morgan from "morgan";
 import path from "path";
 import { fileURLToPath } from "url";
 import appConfig from "./config/app.config.js";
+import corsConfig from "./config/cors.config.js";
 import apiRoutes from "./api/v2/routes/index.routes.js";
 import compression from "compression";
+import { errorHandlerMiddleware } from "./api/v2/middleware/errorHandler.middleware.js";
 
 const app = express();
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 // Middleware
-app.use(cors(appConfig.corsOptions));
+app.use(cors(corsConfig));
 app.use(express.static(path.join(__dirname, appConfig.publicDir)));
 app.use(morgan("combined"));
 app.use(compression());
@@ -22,8 +24,7 @@ app.use("/api/v1", apiRoutes);
 
 // Error handling middleware (should be last)
 app.use((err, req, res, next) => {
-  console.error(err.stack);
-  res.status(500).json({ error: "Something went wrong!" });
+  errorHandlerMiddleware;
 });
 
 // Start server
